Adding user update functionality

+++ src/main/java/com/codeup/adlister/dao/Users.java
@@ -9,5 +9,5 @@ public interface Users {
     Long insert(User user);
     void update(User user);
     void delete(User user);
-    User getUserById(int id);
+    User getUserById(long id);
 }
